Factored out toColorInt ext method to make using in other places easier
