+++ androidshared/src/main/java/org/odk/collect/androidshared/utils/ColorUtils.kt
@@ -1,0 +1,30 @@
+package org.odk.collect.androidshared.utils
+
+import android.graphics.Color
+import androidx.annotation.ColorInt
+
+@ColorInt
+fun String.toColorInt() = try {
+    var sanitizedColor = if (this.startsWith("#")) {
+        this
+    } else {
+        "#$this"
+    }
+
+    if (sanitizedColor.length == 4) {
+        sanitizedColor = shorthandToLonghandHexColor(sanitizedColor)
+    }
+
+    Color.parseColor(sanitizedColor)
+} catch (e: Throwable) {
+    null
+}
+
+private fun shorthandToLonghandHexColor(shorthandColor: String): String {
+    var longHandColor = ""
+    shorthandColor.substring(1).map {
+        longHandColor += it.toString() + it.toString()
+    }
+
+    return "#$longHandColor"
+}
